# openapi.json
{
  "components": {
    "schemas": {
...
      "BarsResponse": {
        "properties": {
          "block": {
            "title": "Block",
            "type": "string"
          },
          "end": {
            "title": "End",
            "type": "string"
          },
          "final": {
            "title": "Final",
            "type": "boolean"
          },
          "interval": {
            "enum": [
              "day",
              "week",
              "month",
              "quarter",
              "year"
            ],
            "title": "Interval",
            "type": "string"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/BarSeriesOut"
            },
            "title": "Series",
            "type": "array"
          },
          "start": {
            "title": "Start",
            "type": "string"
          }
        },
        "required": [
          "interval",
          "block",
          "start",
          "end",
          "final",
          "series"
        ],
        "title": "BarsResponse",
        "type": "object"
      },
...
      "CurveResponse": {
        "properties": {
          "curves": {
            "items": {
              "$ref": "#/components/schemas/CurveOut"
            },
            "title": "Curves",
            "type": "array"
          }
        },
        "required": [
          "curves"
        ],
        "title": "CurveResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The market data platform's API: Treasury CMT yields, curves, spreads and the security master. Instruments by short name; values as decimal strings.",
    "title": "mkt-api",
    "version": "1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/bars": {
      "get": {
        "description": "Bars for a chart: its series at one interval, one fixed block at a time (app/blocks.py).\n\nA series is an instrument (`UST-10Y-CMT`: its yield in percent), `spread(LONG,SHORT)` or\n`fly(WING,BODY,WING)` (basis points). Blocks are a month of days, a year of weeks or a decade of months,\nso a request is a few dozen bars per series and the same block is always the same URL: a finished\nblock (`final`) is cached by the browser for a day. `source` (UST-PAR, H15-TCM) narrows yields to one\npublisher. All arithmetic is Decimal; values are strings.",
        "operationId": "getBars",
        "parameters": [
          {
            "in": "query",
            "name": "series",
            "required": true,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 14,
              "minItems": 1,
              "title": "Series",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": true,
            "schema": {
              "enum": [
                "day",
                "week",
                "month",
                "quarter",
                "year"
              ],
              "title": "Interval",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "block",
            "required": true,
            "schema": {
              "title": "Block",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "",
              "title": "Source",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BarsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Bars"
      }
    },
    "/api/curve": {
      "get": {
        "description": "The curve on a date (default: the latest), plus the same curve `compare` before it (1W, 1M, 1Y).\n\nEach uses the last business day on or before its date that has values.",
        "operationId": "getCurves",
        "parameters": [
          {
            "in": "query",
            "name": "date",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date"
            }
          },
          {
            "in": "query",
            "name": "compare",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "maxItems": 6,
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Compare"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CurveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Curve"
      }
    },
...
    "/api/instruments/{name}": {
      "get": {
        "description": "One instrument by short name or alias, with its identifiers, notes and latest golden value.",
        "operationId": "getInstrument",
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "schema": {
              "title": "Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InstrumentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Instrument"
      }
    },
    "/api/search": {
      "get": {
        "description": "Instruments whose name, alias, identifier or description contains q.",
        "operationId": "searchInstruments",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 60,
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InstrumentSummary"
                  },
                  "title": "Response Searchinstruments",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search"
      }
    },
    "/api/series": {
      "get": {
        "description": "Golden yields (or one source's: UST-PAR, H15-TCM) for instruments over a date range (default: a year).\n\n`interval` sums each series up per week, month, quarter or year: open, high, low and close, so all of\nhistory fits a chart (about 780 monthly bars since 1962, against 16,000 days).",
        "operationId": "getSeries",
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "required": true,
            "schema": {
              "items": {
                "type": "string"
              },
              "maxItems": 14,
              "minItems": 1,
              "title": "Name",
              "type": "array"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "default": "",
              "title": "Source",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "interval",
            "required": false,
            "schema": {
              "default": "day",
              "enum": [
                "day",
                "week",
                "month",
                "quarter",
                "year"
              ],
              "title": "Interval",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Series"
      }
    },
...
  }
}